The finishing touch
I was trying to figure out the math to make the pupils follow the mouse and after a good while, with the help of Mr. Pippin's brilliant code example and bunch of youtube videos, I have managed, at least to make it work.

For the pupils to follow the mouse, I have to code the math out first then all of the pixel inside it. As a newbie in programming, I think I did good this assignment.